resolving macOS CI failure

+++ io/src/json/read_conditions_config.cpp
@@ -55,12 +55,12 @@ traccc::conditions_config read_conditions_config(const nlohmann::json& json) {
         scalar threshold = 0.f;
 
         if (json_binning.contains("shift")) {
-            shift[0] = json_binning["shift"][0].get<double>();
-            shift[1] = json_binning["shift"][1].get<double>();
+            shift[0] = json_binning["shift"][0].get<float>();
+            shift[1] = json_binning["shift"][1].get<float>();
         }
 
         if (json_geom.contains("threshold")) {
-            threshold = json_geom["threshold"].get<double>();
+            threshold = json_geom["threshold"].get<float>();
         }
 
         elements.push_back({geoId, {threshold, shift}});
